Revert "refactor SB* file uploader into top-level HOC"
